requirements.txt
# 核心依赖
torch>=1.8.0
numpy>=1.19.0

# 系统监控
psutil>=5.8.0

# 可选依赖
GPUtil>=1.4.0          # GPU监控支持
PyYAML>=5.4.0          # YAML配置文件支持
pandas>=1.3.0          # 数据分析和导出（可选）

# 开发和测试依赖（可选）
pytest>=6.0.0         # 单元测试
pytest-cov>=2.12.0    # 测试覆盖率
black>=21.0.0          # 代码格式化
flake8>=3.9.0          # 代码检查

# 文档生成（可选）
sphinx>=4.0.0         # 文档生成
sphinx-rtd-theme>=0.5.0  # 文档主题